extracted abstract prior model
