+++ autofit/mapper/__init__.py
@@ -1,5 +1,6 @@
 from autofit.mapper.prior_model.prior_model import PriorModel, CollectionPriorModel, \
-    AbstractPriorModel, AnnotationPriorModel
+    AnnotationPriorModel
+from autofit.mapper.prior_model.abstract_prior_model import AbstractPriorModel
 from .model_mapper import ModelMapper
 from .prior import Prior, UniformPrior, GaussianPrior, Constant, LogUniformPrior, \
     AttributeNameValue
