+++ test/mock.py
@@ -1,6 +1,7 @@
 import inspect
 import typing
 
+import autofit.mapper.prior_model.abstract_prior_model
 import autofit.mapper.prior_model.util
 from autofit.mapper import prior as p
 from autofit.mapper.prior_model import prior_model as pm
@@ -164,7 +165,7 @@ class Redshift(object):
 
 
 # noinspection PyAbstractClass
-class GalaxyModel(pm.AbstractPriorModel):
+class GalaxyModel(autofit.mapper.prior_model.abstract_prior_model.AbstractPriorModel):
     def instance_for_arguments(self, arguments):
         try:
             return Galaxy(redshift=self.redshift.instance_for_arguments(arguments))
@@ -196,7 +197,8 @@ class GalaxyModel(pm.AbstractPriorModel):
     def flat_prior_model_tuples(self):
         return [
             item for item in self.__dict__.items()
-            if isinstance(item[1], pm.AbstractPriorModel)
+            if isinstance(item[1],
+                          autofit.mapper.prior_model.abstract_prior_model.AbstractPriorModel)
         ]
